split build graph dsl spelling
