+++ src/build_graph/lowering/dsl.rs
@@ -1,4 +1,11 @@
-use std::{fmt, str::FromStr};
+#[path = "dsl/fields.rs"]
+mod fields;
+#[path = "dsl/host_effects.rs"]
+mod host_effects;
+#[path = "dsl/idents.rs"]
+mod idents;
+#[path = "dsl/kinds.rs"]
+mod kinds;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(super) enum BuildTargetDslKind {
@@ -35,195 +42,4 @@ pub(super) enum BuildTargetDslIdent {
 pub(super) enum HostEffectResultVariant {
     Ok,
     Err,
-}
-
-impl BuildTargetDslKind {
-    const ALL: &[BuildTargetDslKind] = &[
-        BuildTargetDslKind::Executable,
-        BuildTargetDslKind::Test,
-        BuildTargetDslKind::Library,
-    ];
-    const EXECUTABLE: &'static str = "Executable";
-    const TEST: &'static str = "Test";
-    const LIBRARY: &'static str = "Library";
-
-    pub(super) fn as_str(self) -> &'static str {
-        match self {
-            Self::Executable => Self::EXECUTABLE,
-            Self::Test => Self::TEST,
-            Self::Library => Self::LIBRARY,
-        }
-    }
-
-    pub(super) fn supported_display_list() -> String {
-        let names = Self::ALL
-            .iter()
-            .map(|kind| format!("`{kind}`"))
-            .collect::<Vec<_>>();
-        let Some((last, rest)) = names.split_last() else {
-            return String::new();
-        };
-        if rest.is_empty() {
-            return last.clone();
-        }
-        format!("{}, and {last}", rest.join(", "))
-    }
-}
-
-impl BuildTargetField {
-    const ALL: &[BuildTargetField] = &[
-        BuildTargetField::Name,
-        BuildTargetField::Main,
-        BuildTargetField::Root,
-        BuildTargetField::RootSourceFile,
-        BuildTargetField::OutDir,
-        BuildTargetField::Dependencies,
-        BuildTargetField::Features,
-        BuildTargetField::Exports,
-        BuildTargetField::Packages,
-        BuildTargetField::Link,
-    ];
-    const NAME: &'static str = "name";
-    const MAIN: &'static str = "main";
-    const ROOT: &'static str = "root";
-    const ROOT_SOURCE_FILE: &'static str = "root_source_file";
-    const OUT_DIR: &'static str = "out_dir";
-    const DEPENDENCIES: &'static str = "dependencies";
-    const FEATURES: &'static str = "features";
-    const EXPORTS: &'static str = "exports";
-    const PACKAGES: &'static str = "packages";
-    const LINK: &'static str = "link";
-
-    pub(super) fn as_str(self) -> &'static str {
-        match self {
-            Self::Name => Self::NAME,
-            Self::Main => Self::MAIN,
-            Self::Root => Self::ROOT,
-            Self::RootSourceFile => Self::ROOT_SOURCE_FILE,
-            Self::OutDir => Self::OUT_DIR,
-            Self::Dependencies => Self::DEPENDENCIES,
-            Self::Features => Self::FEATURES,
-            Self::Exports => Self::EXPORTS,
-            Self::Packages => Self::PACKAGES,
-            Self::Link => Self::LINK,
-        }
-    }
-
-    pub(super) fn is_package_link_semantics(self) -> bool {
-        matches!(self, Self::Packages | Self::Link)
-    }
-}
-
-impl FromStr for BuildTargetField {
-    type Err = ();
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        Self::ALL
-            .iter()
-            .copied()
-            .find(|field| field.as_str() == value)
-            .ok_or(())
-    }
-}
-
-impl HostEffectResultVariant {
-    const ALL: &[HostEffectResultVariant] =
-        &[HostEffectResultVariant::Ok, HostEffectResultVariant::Err];
-    const OK: &'static str = "Ok";
-    const ERR: &'static str = "Err";
-
-    pub(super) fn as_str(self) -> &'static str {
-        match self {
-            Self::Ok => Self::OK,
-            Self::Err => Self::ERR,
-        }
-    }
-}
-
-impl fmt::Display for HostEffectResultVariant {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-impl FromStr for HostEffectResultVariant {
-    type Err = ();
-
-    fn from_str(value: &str) -> Result<Self, <Self as FromStr>::Err> {
-        Self::ALL
-            .iter()
-            .copied()
-            .find(|variant| variant.as_str() == value)
-            .ok_or(())
-    }
-}
-
-impl fmt::Display for BuildTargetField {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-impl BuildTargetDslIdent {
-    const ALL: &[BuildTargetDslIdent] = &[
-        BuildTargetDslIdent::Builder,
-        BuildTargetDslIdent::Add,
-        BuildTargetDslIdent::Build,
-        BuildTargetDslIdent::Env,
-        BuildTargetDslIdent::Os,
-        BuildTargetDslIdent::ReadFile,
-    ];
-    const BUILDER: &'static str = "b";
-    const ADD: &'static str = "add";
-    const BUILD: &'static str = "build";
-    const ENV: &'static str = "env";
-    const OS: &'static str = "os";
-    const READ_FILE: &'static str = "read_file";
-
-    pub(super) fn as_str(self) -> &'static str {
-        match self {
-            Self::Builder => Self::BUILDER,
-            Self::Add => Self::ADD,
-            Self::Build => Self::BUILD,
-            Self::Env => Self::ENV,
-            Self::Os => Self::OS,
-            Self::ReadFile => Self::READ_FILE,
-        }
-    }
-}
-
-impl fmt::Display for BuildTargetDslIdent {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-impl FromStr for BuildTargetDslIdent {
-    type Err = ();
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        Self::ALL
-            .iter()
-            .copied()
-            .find(|ident| ident.as_str() == value)
-            .ok_or(())
-    }
-}
-
-impl fmt::Display for BuildTargetDslKind {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-impl FromStr for BuildTargetDslKind {
-    type Err = ();
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        Self::ALL
-            .iter()
-            .copied()
-            .find(|kind| kind.as_str() == value)
-            .ok_or(())
-    }
 }
